requirements.txt
hydra-core==1.3.0
pandas==1.5.1
scikit-learn==1.4.0
xgboost==1.7.1
pytorch-tabnet==3.1.1
pandarallel==1.6.1
requests==2.31.0
tab-transformer-pytorch==0.2.6
matplotlib==3.6.2
seaborn==0.12.1